dartlens_backend/scripts/requirements.txt
requests
pymysql
lxml
python-dotenv